bhavcopy and delivery downloaded moved to single table
Moved both bhavcopy and delivery downloaded to single table, also added
an error_code to the entry - so if you have an older entry with error
code NOT_FOUND it means it was a trading holiday, don't try it again and
get a NOT_FOUND again. 'older' means anything more than 7 days old.