Fixes #5 - We are now using `ResultProxy.first()`
The way ResultProxy.returns_rows and ResultsProxy.rowcount works is
quite unintuitive. ResultProxy.returns_rows only tells whether it's
legal to call fetch* methods. Where as rowcount only reflects the rows
that were attempted to be updated so is always -1 in a `select` query.
Instead now we check for `first`. If exists, it automatically closes the
ResultProxy and returns the row. This is suboptimal but only working
solution I know of.